User CRUD

+++ src/main/java/com/vinhdev97/es/api/service/search/ServiceSearchRequest.java
@@ -16,6 +16,6 @@ import lombok.ToString;
 public class ServiceSearchRequest {
   private Long id;
   private String name;
-  private Long limit;
-  private Long offset;
+  private int limit;
+  private int offset;
 }
