+++ src/main/java/com/vinhdev97/es/api/user/search/UserPaginationSearchResponse.java
@@ -14,12 +14,12 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class UserPaginationSearchResponse {
   private List<UserSearchResponse> items;
-  private Long pageSize;
-  private Long currentPage;
-  private Long totalPage;
+  private int pageSize;
+  private int currentPage;
+  private int totalPage;
 
   public static UserPaginationSearchResponse of(
-      List<UserSearchResponse> items, Long currentPage, Long pageSize, Long totalPage) {
+      List<UserSearchResponse> items, int currentPage, int pageSize, int totalPage) {
     return UserPaginationSearchResponse.builder()
         .items(items)
         .currentPage(currentPage)
